[FLINK-18492][python] Extract the implementation logic of Beam in Operations
